+++ pytest_splunk_addon/standard_lib/app_test_generator.py
@@ -5,8 +5,8 @@ Generates test cases of Fields and CIM.
 """
 import logging
 
-from .fields_tests.fields_test_generator import FieldTestGenerator
-from .cim_tests.cim_test_generator import CIMTestGenerator
+from .fields_tests import FieldTestGenerator
+from .cim_tests import CIMTestGenerator
 
 LOGGER = logging.getLogger("pytest-splunk-addon")
