+++ pytest_splunk_addon/standard_lib/cim_tests/__init__.py
@@ -2,3 +2,10 @@
 """
 Test generation mechanism to verify the CIM compatibility of an Add-on 
 """
+
+from .json_schema import JSONSchema
+from .data_set import DataSet
+from .data_model import DataModel
+from .data_model_handler import DataModelHandler
+from .test_generator import CIMTestGenerator
+from .test_templates import CIMTests
