+++ pytest_splunk_addon/standard_lib/cim_tests/data_model.py
@@ -3,7 +3,7 @@
 Includes DataModel class which handles the DataSets within a data model. 
 """
 
-from .data_set import DataSet
+from . import DataSet
 
 class DataModel(object):
     """
