+++ pytest_splunk_addon/standard_lib/cim_tests/data_set.py
@@ -2,7 +2,7 @@
 """
 Includes DataSet class which handles a single data set
 """
-from ..addon_parser.fields import Field
+from ..addon_parser import Field
 
 class DataSet(object):
     """
