+++ pytest_splunk_addon/standard_lib/cim_tests/test_templates.py
