+++ pytest_splunk_addon/standard_lib/fields_tests/__init__.py
@@ -2,3 +2,7 @@
 """
 Test generation mechanism to verify the field extractions of an Add-on 
 """
+
+from .field_bank import FieldBank
+from .test_generator import FieldTestGenerator
+from .test_templates import FieldTests
