+++ pytest_splunk_addon/standard_lib/fields_tests/field_bank.py
@@ -4,8 +4,8 @@ To enhance the test cases while verifying the field extractions.
 """
 import json
 
-from ..addon_parser.fields import Field
-from ..addon_parser.addon_parser import PropsParser
+from ..addon_parser import Field
+from ..addon_parser import PropsParser
 
 class FieldBank(object):
     """
